# apps/material_agent_service/openapi.yaml
openapi: 3.1.0
info:
  title: Material Agent Service
  description: REST API for VLM-based material assignment to 3D USD objects.
  version: "0.3.0"
servers:
  - url: http://localhost:8000
    description: Local development server
paths:
...
  /pipeline:
    post:
      summary: Start material assignment pipeline
      operationId: startPipeline
      tags: [pipeline]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                usd_file:
                  type: string
                  format: binary
                  description: USD file (required if no session_id)
                session_id:
                  type: string
                  description: Existing session ID from /upload-usd
                user_email:
                  type: string
                  description: Optional user email for usage tracking
                reference_images:
                  type: array
                  items: { type: string, format: binary }
                  description: Reference images for VLM context
                reference_pdfs:
                  type: array
                  items: { type: string, format: binary }
                  description: Reference PDFs to convert to images
                reference_descriptions:
                  type: string
                  description: JSON array of image descriptions
                generated_reference_id:
                  type: string
                  description: Generated reference ID returned by generate-reference-image
                pdf_descriptions:
                  type: string
                  description: JSON array of PDF descriptions
                materials_zip:
                  type: string
                  format: binary
                  description: ZIP with custom materials (materials.yaml + USD library)
                user_prompt:
                  type: string
                  description: Custom VLM prompt
                camera_views:
                  type: string
                  default: "+x+y+z,-x-y-z"
                  description: Comma-separated camera views
                optimize_usd:
                  type: string
                  default: "true"
                  description: Enable USD optimization
                enable_deinstance:
                  type: string
                  default: "true"
                enable_split:
                  type: string
                  default: "true"
                enable_deduplicate:
                  type: string
                  default: "true"
                skip_instances:
                  type: string
                  default: "true"
                skip_prototypes:
                  type: string
                  default: "false"
                skip_existing_materials:
                  type: string
                  default: "false"
                pdf_first_page:
                  type: integer
                  description: First page to convert from PDFs (1-indexed)
                pdf_last_page:
                  type: integer
                  description: Last page to convert from PDFs (1-indexed)
                vlm_model:
                  type: string
                  description: VLM model override
                vlm_max_workers:
                  type: integer
                  description: Max parallel VLM workers (default 64)
      responses:
        "202":
          description: Pipeline started
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionCreated" }
        "400": { description: Invalid input }
        "413": { description: File too large }
  /pipeline/{session_id}/status:
    get:
      summary: Get pipeline execution status
      operationId: getPipelineStatus
      tags: [pipeline]
      parameters:
        - name: session_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Pipeline status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PipelineStatus" }
        "404": { description: Session not found }
...
  /pipeline/{session_id}/results:
    get:
      summary: Get pipeline results (only when completed)
      operationId: getPipelineResults
      tags: [pipeline]
      parameters:
        - name: session_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Pipeline results
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PipelineResults" }
        "202": { description: Still running }
        "404": { description: Session not found }
  /pipeline/{session_id}/cancel:
    post:
      summary: Cancel a running pipeline
      operationId: cancelPipeline
      tags: [pipeline]
      parameters:
        - name: session_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Cancellation acknowledged
          content:
            application/json:
              schema:
                type: object
                properties:
                  session_id: { type: string }
                  status: { type: string, example: cancelling }
                  message: { type: string }
        "400": { description: Cannot cancel (not running) }
        "404": { description: Session not found }
  /pipeline/{session_id}/events:
    get:
      summary: Stream real-time progress events (SSE)
      operationId: streamEvents
      tags: [pipeline]
      parameters:
        - name: session_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Server-Sent Events stream
          content:
            text/event-stream:
              schema:
                type: object
                description: ProgressEvent JSON in SSE data field
                properties:
                  session_id: { type: string }
                  step: { type: string }
                  state: { type: string, enum: [running, completed, failed, cancelled] }
                  percent: { type: integer }
                  message: { type: string }
                  overall_percent: { type: integer }
                  extra: { type: object }
        "404": { description: Session not found }
...
  /artifacts/{session_id}/predictions:
    get:
      summary: Download predictions JSONL
      operationId: downloadPredictions
      tags: [artifacts]
      parameters:
        - name: session_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Predictions JSONL file
          content:
            application/x-ndjson:
              schema: { type: string }
        "404": { description: Predictions not available }
  /artifacts/{session_id}/report:
    get:
      summary: View prediction HTML report
      operationId: viewReport
      tags: [artifacts]
      parameters:
        - name: session_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: HTML prediction report
          content:
            text/html:
              schema: { type: string }
        "404": { description: Not available }
...
  /sessions/{session_id}:
    get:
      summary: Get session details
      operationId: getSession
      tags: [sessions]
      parameters:
        - name: session_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Session metadata
          content:
            application/json:
              schema: { type: object }
        "404": { description: Session not found }
    delete:
      summary: Delete session and all artifacts
      operationId: deleteSession
      tags: [sessions]
      parameters:
        - name: session_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204": { description: Session deleted }
        "404": { description: Session not found }
...
components:
  schemas:
    SessionCreated:
      type: object
      properties:
        session_id: { type: string }
        status: { type: string, example: ready }
        message: { type: string }
        estimated_duration_minutes: { type: integer, nullable: true }
    PipelineStatus:
      type: object
      properties:
        session_id: { type: string }
        status:
          type: string
          enum: [pending, running, completed, failed, cancelled, cancelling]
        current_step:
          type: object
          nullable: true
          properties:
            name: { type: string }
            display_name: { type: string }
            started_at: { type: string, format: date-time }
            progress:
              type: object
              properties:
                current: { type: integer }
                total: { type: integer }
                percent: { type: integer }
                message: { type: string }
            elapsed_seconds: { type: integer }
        completed_steps:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              display_name: { type: string }
              started_at: { type: string, format: date-time }
              completed_at: { type: string, format: date-time }
              duration_seconds: { type: integer }
              stats: { type: object }
        overall_progress:
          type: object
          properties:
            current_step: { type: integer }
            total_steps: { type: integer }
            percent: { type: integer }
            estimated_remaining_seconds: { type: integer, nullable: true }
        preview_images:
          type: array
          items: { type: string }
        can_cancel: { type: boolean }
        elapsed_seconds: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PipelineResults:
      type: object
      properties:
        session_id: { type: string }
        status: { type: string, example: completed }
        stats:
          type: object
          properties:
            original_prim_count: { type: integer }
            prims_processed: { type: integer }
            images_generated: { type: integer }
            predictions_made: { type: integer }
            materials_applied: { type: integer }
        timings:
          type: object
          nullable: true
          properties:
            preparation_seconds: { type: number }
            rendering_total_seconds: { type: number }
            rendering_per_prim_seconds: { type: number }
            prediction_total_seconds: { type: number }
            prediction_per_prim_seconds: { type: number }
            apply_seconds: { type: number }
            total_seconds: { type: number }
        download_urls:
          type: object
          properties:
            output_usd: { type: string }
            predictions: { type: string }
            report: { type: string }
        duration_seconds: { type: integer }
        completed_at: { type: string, format: date-time }
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Optional Bearer token (or set MATERIAL_AGENT_TOKEN env var)
security:
  - {}
  - bearerAuth: []